updated to use more general abstract class for easier sub classing when using non-standard state parsing.
